backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.7.0
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.1.0
django-allauth==0.58.2
django-rest-auth==0.9.5
django-filter==25.1
drf-yasg==1.21.7
channels==4.0.0
channels-redis==4.1.0
pytest-django==4.7.0
python-dotenv==1.0.0
gunicorn==21.2.0
djangorestframework-simplejwt==5.5.0
qrcode==7.4.2 